requirements.txt
# Core scientific computing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine learning and clustering
scikit-learn>=1.0.0
hmmlearn>=0.2.7

# Optimization
cvxpy>=1.2.0
pulp>=2.6.0

# Data APIs
yfinance>=0.1.87
fredapi>=0.5.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Database
sqlite3  # Built-in with Python
sqlalchemy>=1.4.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=0.19.0

# Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.0.0
nbformat>=5.0.0

# Testing
pytest>=6.2.0
pytest-cov>=3.0.0
pytest-mock>=3.6.0

# Code quality
black>=22.0.0
flake8>=4.0.0
mypy>=0.910

# Documentation
sphinx>=4.0.0
sphinx-rtd-theme>=1.0.0

# Development utilities
tqdm>=4.62.0
joblib>=1.1.0